Versioned Storages (#1805) - Launching VS: if user clicks Run Custom button selected tool (if any) should be used as default for launching (including tool's settings like command, instance type etc.)
