+++ client/src/components/pipelines/launch/utilities/versioned-storage-launch-info-equal.js
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2017-2021 EPAM Systems, Inc. (https://www.epam.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+export default function versionedStorageLaunchInfoEqual (info1, info2) {
+  if (!info1 && !info2) {
+    return true;
+  }
+  if (!info1 || !info2) {
+    return false;
+  }
+  const {
+    toolId: toolId1,
+    version: version1
+  } = info1;
+  const {
+    toolId: toolId2,
+    version: version2
+  } = info2;
+  return +toolId1 === +toolId2 && version1 === version2;
+}
